code
This code sets up a web server on port 80 using the ESP32 board and the WiFi library. It defines the WiFi network credentials and creates an instance of the WebServer class. The code sets the LED pin as an output and defines HTML code to be displayed on the web page.The HTML code consists of a header with a title and some CSS styling, a body with a title and two buttons, and a JavaScript function to handle button clicks. The two buttons are used to turn the LED on and off by sending HTTP GET requests to the server.In the setup function, the code connects to the Wi-Fi network, handles the root URL by sending the HTML code, and handles the /on and /off URLs by turning the LED on and off, respectively. It also starts the server.In the loop function, the code continuously listens for incoming client requests and handles them accordingly. This code can be used to control an LED over the internet using a web page.